pyproject.toml
[tool.poetry]
name = "backend-fastapi-template"
version = "0.1.0"
description = "A scalable FastAPI backend template"
authors = ["Usman Khan <work.usmankhan@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.36"
databases = "^0.9.0"
alembic = "^1.14.0"
psycopg = "^3.2.3"
psycopg-binary = "^3.2.3"
pydantic = { extras = ["email"], version = "^2.10.4" }
pydantic-settings = "^2.7.0"
postgres = "^4.0"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
sqlalchemy-utils = "^0.41.2"
pyjwt = "^2.10.1"
httpx = "^0.28.1"
testcontainers = "^4.9.0"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
pytest = "^8.3.4"
ruff = "^0.8.6"
pre-commit = "^4.0.1"

[tool.ruff]
line-length = 88
lint.select = ["E", "F", "W", "C90"]
lint.ignore = ["E501"]